Introduce indirections for retro compatibility
A warning that this is where we are introducing some clutter to force compatibility with the old `MeticaAPI` surface made of static calls and NOT `async`. To do this we wrap async into a coroutine that waits for task completion. Needs more work to stabilize the flow. Some calls were commented out to simplify but should ideally be restored later.