[MET-3642] Storage System (#83)
A storage system with different strategies in place for flexibility and
testability.
This system is not in use yet but will be soon be useful for on-device
data persistence and more.

The `StorageSystem` can be initialized with one of the
`IStorageStrategy` available implementations. Currently we have
1. `FileStorageStrategy`: a simple file based storage with good
compromise between simplicity and case management.
2. `InMemoryStorageStrategy`: basically just a `Dictionary` wrapper but
it can act as a mocker in testing.